requirements.txt
Flask
Flask-Cors
mysqlclient
python-dotenv
gunicorn